Generation Software Bill of Materials (SBOM)
Motivation: In some companies, the development
team has to produce Software Bill of Materials (SBOM)
for their project for compliance reasons:
To track dependencies and licenses across their organisation.
Provide a Module that produces SBOMs
in JSON format.

Changes in the core: Extended the .getArtifact
to return the coursier.Resolution as well.
This is then used to get the license information.

Outside the core: Add a SBOM contrib module
- Generate the most basic CycloneDX SBOM files
  Supporting Java modules for a start
- Provide a basic upload to the Dependency Track server